+++ Assets/_Scripts/Data/LoadDataJson.cs
@@ -94,7 +94,7 @@ public class LoadDataJson : MonoBehaviour
 
                     pM.MoreMine_cost_1 = objJson["Map_Moon"][i]["MoreMine_cost_1"].AsInt;
                     pM.MoreMine_cost_2 = objJson["Map_Moon"][i]["MoreMine_cost_2"].AsInt;
-                    pM.MoreMine_cost_3 = objJson["Map_Moon"][i]["MoreMine_cost_3"].AsInt;
+                    pM.MoreMine_cost_3 = objJson["Map_Moon"][i]["MoreMine_cost_3"].AsFloat;
 
                     pM.Upgrade_time = new List<int>();
                     for (int k = 0; k < objJson["Map_Moon"][i]["Upgrade_time"].Count; k++)
